xmhf-core: change type of vcpu.vmcb_vaddr_ptr from u32 to struct _svm_vmcbfields *; CBMC is pleased and no longer complains about null vmcb dereferences. reports a verification unsuccessful message as expected (we currently violate setprot calling convention)

+++ xmhf/src/xmhf-core/verification/xmhf-core-vdriver.c
@@ -76,7 +76,7 @@ void main() {
 		vcpu.id = 0;													//give a LAPIC id
 		vcpu.cpu_vendor = CPU_VENDOR_AMD;								//stick with AMD now
 		vcpu.npt_vaddr_pts = 0xC8000000;								//where our NPTs reside
-		vcpu.vmcb_vaddr_ptr = (u32)&_xvmcb;								//set vcpu VMCB virtual address to something meaningful
+		vcpu.vmcb_vaddr_ptr = &_xvmcb;								//set vcpu VMCB virtual address to something meaningful
 
 		g_midtable_numentries=1;
